rename JobScheduler to JobDispatcher
